Ignore changes to name attribute of repository resource
A repository's name is also its ID, which means that changing its name
will trigger Terraform to delete the repository and recreate it. By
ignoring changes to the name attribute of the repository, it's possible
to rename a repository via the GitHub UI and then update the
reository's name in Terraform to reflect that.